refactor: Clean up and commenting
Added comment about each vague data coming from the server. The example is put there to make refering and determining to what kind of data is being returned or used in the widget.
Clean up involves removing unneccassary print statements that are not clear and also some commented codes which are not needed and are wrote by me. Also commenting out some exception throwing codes used for debugging purposes.